feat: make all hooks pass
this has required modifications on almost each project

+++ services/fctl/cmd/ledger/transactions/revert.go
@@ -44,7 +44,7 @@ func NewRevertCommand() *cobra.Command {
 				return err
 			}
 
-			rsp, _, err := ledgerClient.TransactionsApi.RevertTransaction(cmd.Context(), ledger, int32(txId)).Execute()
+			rsp, _, err := ledgerClient.TransactionsApi.RevertTransaction(cmd.Context(), ledger, txId).Execute()
 			if err != nil {
 				return errors.Wrapf(err, "reverting transaction")
 			}
